+++ backend/slidy_api/settings.py
@@ -39,7 +39,6 @@ INSTALLED_APPS = [
     # Internal Apps
     'users',
     'slides',
-    'authentication',
     # Third Party Packages
     'rest_framework',
     'rest_framework.authtoken',
@@ -134,7 +133,7 @@ DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 REST_FRAMEWORK = {
    "DEFAULT_AUTHENTICATION_CLASSES": [
         "knox.auth.TokenAuthentication",
-        "authentication.authentication.TokenAuthentication",
+        "users.authentication.TokenAuthentication",
    ],
    "DEFAULT_PERMISSION_CLASSES": [
         "rest_framework.permissions.IsAuthenticatedOrReadOnly"
