+++ backend/slidy_api/urls.py
@@ -3,6 +3,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/v1/auth/', include('authentication.urls')),
+    path('api/v1/auth/', include('users.urls')),
     path('api/v1/', include('slidy_api.routers'))
 ]
